refactor: Endpoint 이름으로 API 파일명 변경
