+++ src/components/editpage/NoteEditor.tsx
@@ -13,21 +13,23 @@ type NoteEditorProps = {
 const NoteEditor = ({ keywordList, memo, onMemoChanged }: NoteEditorProps) => {
   return (
     <View style={styles.noteEditorContainer}>
-      <Keywords keywordList={keywordList} />
-      <TextInput
-        style={styles.noteEditor}
-        multiline={true}
-        numberOfLines={3}
-        maxLength={100}
-        placeholder={
-          "오늘 나만의 온도는 어땠나요?\n오늘의 하루를 컬러와 간단한 문장으로 표현해보세요."
-        }
-        placeholderTextColor={Colors.black40}
-        autoFocus
-        onChangeText={(memo) => {
-          onMemoChanged(memo);
-        }}
-      />
+      <View style={{ flex: 1 }}>
+        <Keywords keywordList={keywordList} />
+        <TextInput
+          style={styles.noteEditor}
+          multiline={true}
+          numberOfLines={3}
+          maxLength={100}
+          placeholder={
+            "오늘 나만의 온도는 어땠나요?\n오늘의 하루를 컬러와 간단한 문장으로 표현해보세요."
+          }
+          placeholderTextColor={Colors.black40}
+          autoFocus
+          onChangeText={(memo) => {
+            onMemoChanged(memo);
+          }}
+        />
+      </View>
       <View style={{ flexDirection: "row", justifyContent: "flex-end" }}>
         <Text style={styles.noteCountingLabel}>{memo.length}</Text>
         <Text style={styles.noteMaxLabel}> /100</Text>
@@ -40,6 +42,7 @@ export default NoteEditor;
 
 const styles = StyleSheet.create({
   noteEditorContainer: {
+    flex: 1,
     paddingVertical: 24,
     paddingHorizontal: 16,
     gap: 16,
@@ -50,6 +53,7 @@ const styles = StyleSheet.create({
     ...typography.body,
     textAlignVertical: "top",
     minHeight: 64,
+    // maxHeight: 80,
   },
   noteCountingLabel: { ...typography.label1, color: Colors.black100 },
   noteMaxLabel: { ...typography.label1, color: Colors.black40 },
